refactor(kernel,protocol): split large files into focused modules (#304)
Phase 1 of File Splitting Epic - split 3 low-risk files:

- input.rs (1,348 lines) -> 6 modules: key, mouse, terminal, session, health
- option.rs (1,487 lines) -> 6 modules: value, scope, constraint, error, spec
- motion.rs (1,178 lines) -> 3 modules: types, engine

Total: 4,013 lines reorganized into 15 focused modules.
External APIs unchanged via flat re-exports.
All tests passing.

Parts of #304
Refs #300, #301, #303
