+++ lib/kernel/src/core/option/mod.rs
@@ -40,614 +40,24 @@
 //! let value = registry.get("number", OptionScopeId::Global);
 //! ```
 
-use std::{borrow::Cow, collections::HashMap, fmt};
+mod constraint;
+mod error;
+mod scope;
+mod spec;
+mod value;
+
+// Re-export all types for API compatibility
+pub use constraint::{ConstraintError, OptionConstraint};
+pub use error::{OptionError, SetResult};
+pub use scope::{OptionScope, OptionScopeId};
+pub use spec::OptionSpec;
+pub use value::OptionValue;
+
+use std::collections::HashMap;
 
 use reovim_arch::sync::RwLock;
 
 use crate::mm::{BufferId, WindowId};
-
-// ============================================================================
-// OptionValue - Type-safe option values
-// ============================================================================
-
-/// Type-safe option value.
-///
-/// This is the core value type for editor options. Each variant maps to
-/// a common configuration pattern:
-/// - `Bool`: Toggle settings (number, wrap, etc.)
-/// - `Integer`: Numeric settings (tabwidth, scrolloff)
-/// - `String`: Free-form text (theme name, paths)
-/// - `Choice`: Enum-like selection from predefined values
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum OptionValue {
-    /// Boolean on/off option (e.g., `number`, `relativenumber`)
-    Bool(bool),
-
-    /// Integer option (e.g., `tabwidth`, `scrolloff`)
-    Integer(i64),
-
-    /// String option (e.g., `theme`, `signcolumn`)
-    String(String),
-
-    /// Choice option with predefined valid values.
-    ///
-    /// The `choices` field stores all valid values for validation.
-    Choice {
-        /// Current selected value
-        value: String,
-        /// All valid choices (for validation and completion)
-        choices: Vec<String>,
-    },
-}
-
-impl OptionValue {
-    // ========================================================================
-    // Constructors
-    // ========================================================================
-
-    /// Create a boolean value.
-    #[must_use]
-    pub const fn bool(value: bool) -> Self {
-        Self::Bool(value)
-    }
-
-    /// Create an integer value.
-    #[must_use]
-    pub const fn int(value: i64) -> Self {
-        Self::Integer(value)
-    }
-
-    /// Create a string value.
-    #[must_use]
-    pub fn string(value: impl Into<String>) -> Self {
-        Self::String(value.into())
-    }
-
-    /// Create a choice value.
-    ///
-    /// # Panics
-    ///
-    /// Panics if `choices` is empty or if `value` is not in `choices`.
-    #[must_use]
-    pub fn choice(value: impl Into<String>, choices: Vec<String>) -> Self {
-        let value = value.into();
-        debug_assert!(!choices.is_empty(), "choices must not be empty");
-        debug_assert!(choices.contains(&value), "value must be in choices");
-        Self::Choice { value, choices }
-    }
-
-    // ========================================================================
-    // Accessors
-    // ========================================================================
-
-    /// Try to get as boolean.
-    #[must_use]
-    pub const fn as_bool(&self) -> Option<bool> {
-        match self {
-            Self::Bool(b) => Some(*b),
-            _ => None,
-        }
-    }
-
-    /// Try to get as integer.
-    #[must_use]
-    pub const fn as_int(&self) -> Option<i64> {
-        match self {
-            Self::Integer(i) => Some(*i),
-            _ => None,
-        }
-    }
-
-    /// Try to get as string reference.
-    #[must_use]
-    pub fn as_str(&self) -> Option<&str> {
-        match self {
-            Self::String(s) => Some(s),
-            Self::Choice { value, .. } => Some(value),
-            _ => None,
-        }
-    }
-
-    /// Get the type name for error messages.
-    #[must_use]
-    pub const fn type_name(&self) -> &'static str {
-        match self {
-            Self::Bool(_) => "bool",
-            Self::Integer(_) => "integer",
-            Self::String(_) => "string",
-            Self::Choice { .. } => "choice",
-        }
-    }
-
-    /// Check if this value has the same type as another.
-    #[must_use]
-    pub const fn same_type(&self, other: &Self) -> bool {
-        matches!(
-            (self, other),
-            (Self::Bool(_), Self::Bool(_))
-                | (Self::Integer(_), Self::Integer(_))
-                | (Self::String(_), Self::String(_))
-                | (Self::Choice { .. }, Self::Choice { .. })
-        )
-    }
-}
-
-impl fmt::Display for OptionValue {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Bool(b) => write!(f, "{b}"),
-            Self::Integer(i) => write!(f, "{i}"),
-            Self::String(s) => write!(f, "{s}"),
-            Self::Choice { value, .. } => write!(f, "{value}"),
-        }
-    }
-}
-
-// ============================================================================
-// OptionScope - Where an option applies
-// ============================================================================
-
-/// Scope where an option applies.
-///
-/// Determines the granularity of option storage:
-/// - `Global`: Single value for the entire editor
-/// - `Buffer`: Per-buffer values (e.g., `tabwidth`)
-/// - `Window`: Per-window values (e.g., `number`)
-///
-/// Window-scoped options can also have buffer-local defaults.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum OptionScope {
-    /// Applies globally to the entire editor.
-    #[default]
-    Global,
-
-    /// Per-buffer setting (e.g., `filetype`, `tabwidth`, `expandtab`).
-    Buffer,
-
-    /// Per-window setting (e.g., `number`, `relativenumber`, `wrap`).
-    Window,
-}
-
-impl OptionScope {
-    /// Get display name for this scope.
-    #[must_use]
-    pub const fn display_name(self) -> &'static str {
-        match self {
-            Self::Global => "global",
-            Self::Buffer => "buffer",
-            Self::Window => "window",
-        }
-    }
-}
-
-impl fmt::Display for OptionScope {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.display_name())
-    }
-}
-
-// ============================================================================
-// OptionScopeId - Runtime scope identifier
-// ============================================================================
-
-/// Runtime scope identifier for option access.
-///
-/// Used when getting or setting option values to specify the exact
-/// scope context (which buffer, which window, or global).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
-pub enum OptionScopeId {
-    /// Global scope (no buffer/window context).
-    #[default]
-    Global,
-    /// Buffer-local scope with specific buffer ID.
-    Buffer(BufferId),
-    /// Window-local scope with specific window ID.
-    Window(WindowId),
-}
-
-impl fmt::Display for OptionScopeId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Global => write!(f, "global"),
-            Self::Buffer(id) => write!(f, "buffer({id:?})"),
-            Self::Window(id) => write!(f, "window({id:?})"),
-        }
-    }
-}
-
-// ============================================================================
-// OptionConstraint - Validation constraints
-// ============================================================================
-
-/// Constraints for option value validation.
-///
-/// These constraints are checked when setting option values.
-/// Not all constraints apply to all value types.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct OptionConstraint {
-    /// For Integer: minimum value (inclusive).
-    pub min: Option<i64>,
-    /// For Integer: maximum value (inclusive).
-    pub max: Option<i64>,
-    /// For String: minimum length.
-    pub min_length: Option<usize>,
-    /// For String: maximum length.
-    pub max_length: Option<usize>,
-}
-
-impl OptionConstraint {
-    /// Create a constraint with no restrictions.
-    #[must_use]
-    pub const fn none() -> Self {
-        Self {
-            min: None,
-            max: None,
-            min_length: None,
-            max_length: None,
-        }
-    }
-
-    /// Create a constraint with integer range (inclusive).
-    #[must_use]
-    pub const fn range(min: i64, max: i64) -> Self {
-        Self {
-            min: Some(min),
-            max: Some(max),
-            min_length: None,
-            max_length: None,
-        }
-    }
-
-    /// Create a constraint with minimum integer value.
-    #[must_use]
-    pub const fn min(min: i64) -> Self {
-        Self {
-            min: Some(min),
-            max: None,
-            min_length: None,
-            max_length: None,
-        }
-    }
-
-    /// Create a constraint with maximum integer value.
-    #[must_use]
-    pub const fn max(max: i64) -> Self {
-        Self {
-            min: None,
-            max: Some(max),
-            min_length: None,
-            max_length: None,
-        }
-    }
-
-    /// Create a constraint with string length range.
-    #[must_use]
-    pub const fn string_length(min_length: usize, max_length: usize) -> Self {
-        Self {
-            min: None,
-            max: None,
-            min_length: Some(min_length),
-            max_length: Some(max_length),
-        }
-    }
-
-    /// Validate a value against this constraint.
-    ///
-    /// # Errors
-    ///
-    /// Returns `ConstraintError` if the value violates the constraint.
-    pub fn validate(&self, value: &OptionValue) -> Result<(), ConstraintError> {
-        match value {
-            OptionValue::Integer(i) => {
-                if let Some(min) = self.min
-                    && *i < min
-                {
-                    return Err(ConstraintError::BelowMinimum { value: *i, min });
-                }
-                if let Some(max) = self.max
-                    && *i > max
-                {
-                    return Err(ConstraintError::AboveMaximum { value: *i, max });
-                }
-            }
-            OptionValue::String(s) => {
-                if let Some(min_len) = self.min_length
-                    && s.len() < min_len
-                {
-                    return Err(ConstraintError::StringTooShort {
-                        len: s.len(),
-                        min: min_len,
-                    });
-                }
-                if let Some(max_len) = self.max_length
-                    && s.len() > max_len
-                {
-                    return Err(ConstraintError::StringTooLong {
-                        len: s.len(),
-                        max: max_len,
-                    });
-                }
-            }
-            OptionValue::Choice { value, choices } => {
-                if !choices.contains(value) {
-                    return Err(ConstraintError::InvalidChoice {
-                        value: value.clone(),
-                        choices: choices.clone(),
-                    });
-                }
-            }
-            OptionValue::Bool(_) => {
-                // No constraints for boolean values
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Constraint validation error.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ConstraintError {
-    /// Integer value is below minimum.
-    BelowMinimum {
-        /// The actual value
-        value: i64,
-        /// The minimum allowed
-        min: i64,
-    },
-    /// Integer value is above maximum.
-    AboveMaximum {
-        /// The actual value
-        value: i64,
-        /// The maximum allowed
-        max: i64,
-    },
-    /// String is too short.
-    StringTooShort {
-        /// Actual length
-        len: usize,
-        /// Minimum required length
-        min: usize,
-    },
-    /// String is too long.
-    StringTooLong {
-        /// Actual length
-        len: usize,
-        /// Maximum allowed length
-        max: usize,
-    },
-    /// Choice value is not in the allowed choices.
-    InvalidChoice {
-        /// The invalid value
-        value: String,
-        /// The valid choices
-        choices: Vec<String>,
-    },
-}
-
-impl fmt::Display for ConstraintError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::BelowMinimum { value, min } => {
-                write!(f, "value {value} is below minimum {min}")
-            }
-            Self::AboveMaximum { value, max } => {
-                write!(f, "value {value} is above maximum {max}")
-            }
-            Self::StringTooShort { len, min } => {
-                write!(f, "string length {len} is below minimum {min}")
-            }
-            Self::StringTooLong { len, max } => {
-                write!(f, "string length {len} is above maximum {max}")
-            }
-            Self::InvalidChoice { value, choices } => {
-                write!(f, "'{value}' is not a valid choice (valid: {choices:?})")
-            }
-        }
-    }
-}
-
-impl std::error::Error for ConstraintError {}
-
-// ============================================================================
-// OptionSpec - Complete option specification
-// ============================================================================
-
-/// Complete specification for an editor option.
-///
-/// This is the **MECHANISM** type - it defines the structure of options.
-/// **POLICY** (which options exist and what they do) is in modules.
-///
-/// # Example
-///
-/// ```ignore
-/// use reovim_kernel::api::v1::*;
-///
-/// let spec = OptionSpec::new("tabwidth", "Tab width in spaces", OptionValue::int(4))
-///     .with_short("tw")
-///     .with_constraint(OptionConstraint::range(1, 32))
-///     .with_scope(OptionScope::Buffer);
-/// ```
-#[derive(Debug, Clone)]
-pub struct OptionSpec {
-    /// Full option name (e.g., "number", "tabwidth").
-    pub name: Cow<'static, str>,
-    /// Short alias (e.g., "nu" for "number").
-    pub short_form: Option<Cow<'static, str>>,
-    /// Human-readable description.
-    pub description: Cow<'static, str>,
-    /// Default value (also defines the type).
-    pub default: OptionValue,
-    /// Constraints for validation.
-    pub constraint: OptionConstraint,
-    /// Scope (global, buffer-local, window-local).
-    pub scope: OptionScope,
-}
-
-impl OptionSpec {
-    /// Create a new option specification.
-    #[must_use]
-    pub fn new(
-        name: impl Into<Cow<'static, str>>,
-        description: impl Into<Cow<'static, str>>,
-        default: OptionValue,
-    ) -> Self {
-        Self {
-            name: name.into(),
-            description: description.into(),
-            default,
-            short_form: None,
-            constraint: OptionConstraint::none(),
-            scope: OptionScope::default(),
-        }
-    }
-
-    /// Add a short alias.
-    #[must_use]
-    pub fn with_short(mut self, short: impl Into<Cow<'static, str>>) -> Self {
-        self.short_form = Some(short.into());
-        self
-    }
-
-    /// Add validation constraints.
-    #[must_use]
-    pub const fn with_constraint(mut self, constraint: OptionConstraint) -> Self {
-        self.constraint = constraint;
-        self
-    }
-
-    /// Set the scope.
-    #[must_use]
-    pub const fn with_scope(mut self, scope: OptionScope) -> Self {
-        self.scope = scope;
-        self
-    }
-
-    /// Check if this option matches a name (full name or alias).
-    #[must_use]
-    pub fn matches_name(&self, query: &str) -> bool {
-        self.name == query
-            || self
-                .short_form
-                .as_ref()
-                .is_some_and(|s: &Cow<'static, str>| s.as_ref() == query)
-    }
-
-    /// Validate a value against this option's type and constraints.
-    ///
-    /// # Errors
-    ///
-    /// Returns error if value type doesn't match or violates constraints.
-    pub fn validate(&self, value: &OptionValue) -> Result<(), OptionError> {
-        // Type check
-        if !self.default.same_type(value) {
-            return Err(OptionError::TypeMismatch {
-                name: self.name.to_string(),
-                expected: self.default.type_name(),
-                got: value.type_name(),
-            });
-        }
-
-        // Constraint check
-        self.constraint
-            .validate(value)
-            .map_err(|e| OptionError::ValidationFailed {
-                name: self.name.to_string(),
-                reason: e.to_string(),
-            })
-    }
-}
-
-// ============================================================================
-// OptionError - Error type for option operations
-// ============================================================================
-
-/// Error type for option operations.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum OptionError {
-    /// Option not found in registry.
-    NotFound(String),
-
-    /// Option already registered.
-    AlreadyExists(String),
-
-    /// Validation failed.
-    ValidationFailed {
-        /// Option name
-        name: String,
-        /// Reason for failure
-        reason: String,
-    },
-
-    /// Type mismatch.
-    TypeMismatch {
-        /// Option name
-        name: String,
-        /// Expected type
-        expected: &'static str,
-        /// Actual type
-        got: &'static str,
-    },
-
-    /// Alias conflicts with existing name.
-    AliasConflict(String),
-
-    /// Scope mismatch (e.g., setting buffer-local for global-only option).
-    ScopeMismatch {
-        /// Option name
-        name: String,
-        /// Option's declared scope
-        option_scope: OptionScope,
-        /// Requested scope
-        requested: OptionScopeId,
-    },
-}
-
-impl fmt::Display for OptionError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::NotFound(name) => write!(f, "option not found: {name}"),
-            Self::AlreadyExists(name) => write!(f, "option already exists: {name}"),
-            Self::ValidationFailed { name, reason } => {
-                write!(f, "validation failed for '{name}': {reason}")
-            }
-            Self::TypeMismatch {
-                name,
-                expected,
-                got,
-            } => {
-                write!(f, "type mismatch for '{name}': expected {expected}, got {got}")
-            }
-            Self::AliasConflict(alias) => {
-                write!(f, "alias conflicts with existing name: {alias}")
-            }
-            Self::ScopeMismatch {
-                name,
-                option_scope,
-                requested,
-            } => {
-                write!(
-                    f,
-                    "scope mismatch for '{name}': option is {option_scope}, requested {requested}"
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for OptionError {}
-
-// ============================================================================
-// SetResult - Result of a set operation
-// ============================================================================
-
-/// Result of a successful set operation.
-#[derive(Debug, Clone)]
-pub struct SetResult {
-    /// Previous value (None if was default).
-    pub old_value: Option<OptionValue>,
-    /// New value that was set.
-    pub new_value: OptionValue,
-}
 
 // ============================================================================
 // OptionRegistry - Thread-safe option storage
@@ -1081,139 +491,9 @@ impl OptionRegistry {
     }
 }
 
-// ============================================================================
-// Tests
-// ============================================================================
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // ========================================================================
-    // OptionValue tests
-    // ========================================================================
-
-    #[test]
-    fn test_option_value_bool() {
-        let value = OptionValue::bool(true);
-        assert_eq!(value.as_bool(), Some(true));
-        assert_eq!(value.as_int(), None);
-        assert_eq!(value.as_str(), None);
-        assert_eq!(value.type_name(), "bool");
-    }
-
-    #[test]
-    fn test_option_value_int() {
-        let value = OptionValue::int(42);
-        assert_eq!(value.as_int(), Some(42));
-        assert_eq!(value.as_bool(), None);
-        assert_eq!(value.type_name(), "integer");
-    }
-
-    #[test]
-    fn test_option_value_string() {
-        let value = OptionValue::string("hello");
-        assert_eq!(value.as_str(), Some("hello"));
-        assert_eq!(value.as_bool(), None);
-        assert_eq!(value.type_name(), "string");
-    }
-
-    #[test]
-    fn test_option_value_choice() {
-        let value = OptionValue::choice("dark", vec!["light".into(), "dark".into()]);
-        assert_eq!(value.as_str(), Some("dark"));
-        assert_eq!(value.type_name(), "choice");
-    }
-
-    #[test]
-    fn test_option_value_same_type() {
-        assert!(OptionValue::bool(true).same_type(&OptionValue::bool(false)));
-        assert!(OptionValue::int(1).same_type(&OptionValue::int(2)));
-        assert!(!OptionValue::bool(true).same_type(&OptionValue::int(1)));
-    }
-
-    #[test]
-    fn test_option_value_display() {
-        assert_eq!(OptionValue::bool(true).to_string(), "true");
-        assert_eq!(OptionValue::int(42).to_string(), "42");
-        assert_eq!(OptionValue::string("hello").to_string(), "hello");
-    }
-
-    // ========================================================================
-    // OptionConstraint tests
-    // ========================================================================
-
-    #[test]
-    fn test_constraint_range() {
-        let constraint = OptionConstraint::range(1, 10);
-
-        // Valid values
-        assert!(constraint.validate(&OptionValue::int(1)).is_ok());
-        assert!(constraint.validate(&OptionValue::int(5)).is_ok());
-        assert!(constraint.validate(&OptionValue::int(10)).is_ok());
-
-        // Invalid values
-        assert!(constraint.validate(&OptionValue::int(0)).is_err());
-        assert!(constraint.validate(&OptionValue::int(11)).is_err());
-    }
-
-    #[test]
-    fn test_constraint_string_length() {
-        let constraint = OptionConstraint::string_length(2, 5);
-
-        assert!(constraint.validate(&OptionValue::string("ab")).is_ok());
-        assert!(constraint.validate(&OptionValue::string("abcde")).is_ok());
-        assert!(constraint.validate(&OptionValue::string("a")).is_err());
-        assert!(constraint.validate(&OptionValue::string("abcdef")).is_err());
-    }
-
-    #[test]
-    fn test_constraint_choice() {
-        let value = OptionValue::choice("valid", vec!["valid".into(), "other".into()]);
-        let constraint = OptionConstraint::none();
-        assert!(constraint.validate(&value).is_ok());
-    }
-
-    // ========================================================================
-    // OptionSpec tests
-    // ========================================================================
-
-    #[test]
-    fn test_option_spec_builder() {
-        let spec = OptionSpec::new("number", "Show line numbers", OptionValue::bool(false))
-            .with_short("nu")
-            .with_scope(OptionScope::Window);
-
-        assert_eq!(spec.name, "number");
-        assert_eq!(spec.short_form.as_deref(), Some("nu"));
-        assert_eq!(spec.scope, OptionScope::Window);
-    }
-
-    #[test]
-    fn test_option_spec_matches_name() {
-        let spec = OptionSpec::new("number", "desc", OptionValue::bool(false)).with_short("nu");
-
-        assert!(spec.matches_name("number"));
-        assert!(spec.matches_name("nu"));
-        assert!(!spec.matches_name("other"));
-    }
-
-    #[test]
-    fn test_option_spec_validate() {
-        let spec = OptionSpec::new("tabwidth", "Tab width", OptionValue::int(4))
-            .with_constraint(OptionConstraint::range(1, 32));
-
-        assert!(spec.validate(&OptionValue::int(4)).is_ok());
-        assert!(spec.validate(&OptionValue::int(1)).is_ok());
-        assert!(spec.validate(&OptionValue::int(32)).is_ok());
-        assert!(spec.validate(&OptionValue::int(0)).is_err());
-        assert!(spec.validate(&OptionValue::int(33)).is_err());
-        assert!(spec.validate(&OptionValue::bool(true)).is_err()); // Type mismatch
-    }
-
-    // ========================================================================
-    // OptionRegistry tests
-    // ========================================================================
 
     #[test]
     fn test_registry_register() {
